backend/gateway/docter_service/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
python-dotenv
Flask-JWT-Extended